+++ noodles-bam/src/lazy/record/data.rs
@@ -100,8 +100,12 @@ mod tests {
 
         let data = Data::new(&[b'N', b'H', b'C', 0x01]);
         let actual: Vec<_> = data.iter().collect::<io::Result<_>>()?;
-        let expected = [([b'N', b'H'], Value::UInt8(1))];
-        assert_eq!(actual, expected);
+
+        assert_eq!(actual.len(), 1);
+
+        let (actual_tag, actual_value) = &actual[0];
+        assert_eq!(actual_tag, &[b'N', b'H']);
+        assert!(matches!(actual_value, Value::UInt8(1)));
 
         Ok(())
     }